being conscientious about datatype widths: use uint8_t* instead of char* and make use of uint_fast*_t types

+++ src/shmem/include/ptl_internal_LE.h
@@ -4,9 +4,9 @@
 #include "ptl_internal_PT.h"
 #include "ptl_internal_commpad.h"
 
-void INTERNAL      PtlInternalLENISetup(unsigned int ni,
-                                        ptl_size_t   limit);
-void INTERNAL      PtlInternalLENITeardown(unsigned int ni);
+void INTERNAL PtlInternalLENISetup(uint_fast8_t ni,
+                                   ptl_size_t   limit);
+void INTERNAL      PtlInternalLENITeardown(uint_fast8_t ni);
 ptl_pid_t INTERNAL PtlInternalLEDeliver(ptl_table_entry_t *restrict     t,
                                         ptl_internal_header_t *restrict h);
